+++ src/js/features/deck-view-features/index.ts
@@ -1,5 +1,6 @@
 import DeckDisplay from "./deck-display";
+import PriceOptions from "./price-options";
 import TokenList from "./token-list";
 import "./index.css";
 
-export default [DeckDisplay, TokenList];
+export default [DeckDisplay, PriceOptions, TokenList];