+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Cleanup Improver
   - Add feature to keep deck organized by specific settings
   - Add feature to include headings in deck organization
+- Price Options
+  - Add feature to display no price on deck view page (#68)
 
 # 0.7.0 - Pre-release